+++ fluent.runtime/CHANGELOG.rst
@@ -8,6 +8,7 @@ fluent.runtime next
   ``fluent.runtime.FluentBundle.add_resource``.
 * Removed ``fluent.runtime.FluentBundle.add_messages``.
 * Replaced ``bundle.format()`` with ``bundle.format_pattern(bundle.get_message().value)``.
+* Added ``fluent.runtime.FluentLocalization`` as main entrypoint for applications.
 
 fluent.runtime 0.2 (September 10, 2019)
 ---------------------------------------